Bugs fixed in Routine Microservice.

+++ src/Microservices/RoutineMicroservice/RoutineMS_Infraestructure/Repositories/RoutineRepository.cs
@@ -4,8 +4,8 @@ using RoutineMS_Core.Exceptions;
 using RoutineMS_Core.Models.Entities;
 using RoutineMS_Core.Repositories;
 using RoutineMS_Infraestructure.Data;
-using RoutineMS_Infraestructure.Extensions;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -15,23 +15,46 @@ namespace RoutineMS_Infraestructure.Repositories
     {
         public RoutineRepository(RoutineMsDbContext context) : base(context) { }
 
-        public async Task<ExerciseToDoDto> GetExerciseToDoFromRoutineAsync(Guid userId)
+        public override async Task<Routine> GetByIdAsync(Guid id)
         {
             try
             {
-                int day = DateTime.UtcNow.GetDayOfWeek();
+                var routine = await _context.Routines.Include(r => r.Exercise).ThenInclude(e => e.SetDetail).Where(r => r.ExerciseId == id).FirstOrDefaultAsync();
 
+                if (routine is null) throw new RoutineMSException($"Routine with id {id} could not be found") { StatusCode = 404};
+                
+                return routine;
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+        }
+        public async Task<ExerciseToDoDto> GetExerciseToDoFromRoutineAsync(Guid userId, int day)
+        {
+            try
+            {
                 var routinesForToday = await
                     _context
                     .Routines
                     .AsNoTrackingWithIdentityResolution()
                     .Include(r => r.Days).AsSplitQuery()
                     .Include(r => r.Exercise).ThenInclude(e => e.SetDetail)
-                    .Where(r => (r.UserId == userId) && (r.Days.Any(d => d.Id == day)))
-                    .OrderBy(r => r.Exercise.CategoryName)
+                    .OrderBy(r => r.Exercise.CategoryName).ThenBy(r => r.Exercise.Name)
                     .ToListAsync();
 
-                if (routinesForToday is null || routinesForToday?.Count == 0) return null;
+                List<Routine> routines = new();
+
+                foreach (var r in routinesForToday)
+                {
+                    if (!(r.UserId == userId)) continue;
+
+                    if (!(r.Days.Any(d => d.Id == day))) continue;
+
+                    routines.Add(r);
+                }
+
+                if(routines.Count == 0) return null;
 
                 Exercise exercise = null;
 
@@ -63,23 +86,39 @@ namespace RoutineMS_Infraestructure.Repositories
             }
 
         }
-
-        public async Task<bool> RemoveRoutineByExercise(Guid exerciseId)
+ 
+        public async Task<bool> RemoveRoutineByExerciseAsync(Guid exerciseId, Guid userId)
         {
             try
             {
                 Routine routine = await _context
                     .Routines
                     .Include(r => r.Exercise).ThenInclude(e => e.SetDetail)
-                    .FirstOrDefaultAsync(r => r.ExerciseId == exerciseId);
+                    .FirstOrDefaultAsync(r => r.ExerciseId == exerciseId && r.UserId == userId);
 
                 if (routine is null) throw new RoutineMSException("Invalid exercise to remove!") { StatusCode = 404};
 
                 _context.Routines.Remove(routine);
-                _context.Exercises.Remove(routine.Exercise);
-                _context.SetsDetails.Remove(routine.Exercise.SetDetail);
+                if(routine.Exercise is not null)_context.Exercises.Remove(routine.Exercise);
+                if (routine.Exercise.SetDetail is not null) _context.SetsDetails.Remove(routine.Exercise.SetDetail);
 
                 return true;
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+        }
+
+        public async Task<bool> TheExerciseIsInRoutineAsync(Guid exerciseId, Guid userId)
+        {
+            try
+            {
+                Routine routine = await _context
+                    .Routines
+                    .FirstOrDefaultAsync(r => r.ExerciseId == exerciseId && r.UserId == userId);
+
+                return (routine is not null) ?true:false;
             }
             catch (Exception)
             {
